WithSpanObserver added
