+++ stmt.go
@@ -35,7 +35,7 @@ func (s *stmt) Query(args []driver.Value) (driver.Rows, error) {
 func (s *stmt) ExecContext(ctx context.Context, query string, args []driver.NamedValue) (driver.Result, error) {
 	span := s.tracer.newSpan(ctx)
 	if s.tracer.saveQuery {
-		span.SetTag(TagQuery, query)
+		span.SetTag(spanTagQuery, query)
 	}
 	defer span.Finish()
 	if execerContext, ok := s.stmt.(driver.ExecerContext); ok {
@@ -52,7 +52,7 @@ func (s *stmt) ExecContext(ctx context.Context, query string, args []driver.Name
 func (s *stmt) QueryContext(ctx context.Context, query string, args []driver.NamedValue) (rows driver.Rows, err error) {
 	span := s.tracer.newSpan(ctx)
 	if s.tracer.saveQuery {
-		span.SetTag(TagQuery, query)
+		span.SetTag(spanTagQuery, query)
 	}
 	defer span.Finish()
 	if queryerContext, ok := s.stmt.(driver.QueryerContext); ok {
